fix(jit:175972df): drop redundant u64 cast in enumeration test
Q is already a u64 const parameter, so (Q as u64).pow(..) trips
clippy::unnecessary_cast under -D warnings.

+++ crates/gf2-algebra/tests/permanental_rank.rs
@@ -106,7 +106,7 @@ fn assert_exhaustive_agreement<const Q: u64>(n: usize, k: usize) {
                 // Every matrix was visited exactly once.
                 assert_eq!(
                     examined,
-                    (Q as u64).pow(cells as u32),
+                    Q.pow(cells as u32),
                     "q={Q} n={n} k={k}: enumerated {examined} matrices, expected q^(n*k)"
                 );
                 // A deficient matrix exists at every shape tested here (the
